Implement Auto Login
After RECIEVE_ENABLE is returned, check if the localstorage name
is set. If it is, then use that name and the first wallet address
to try to login. With Authenticate, a 3rd parameter for noRedirect
is added to tell the method not to redirect on success or fail.

Tested with metamask and nifty