Bump web3 version
The previous version was not compatible with wallet connect and was showing the error `this.send is not a function`. This behavior is documented in this issue here: https://github.com/ethereum/web3.js/issues/3790. The reccomendation was to downgrade to 1.2.x but in this case, we are upgrading to it.